configure db to api
