+++ components/databases/focalboard/src/mutator.ts
@@ -7,7 +7,7 @@ import {BoardView, ISortOption, createBoardView, KanbanCalculationFields} from '
 import {Card, createCard} from './blocks/card'
 import {FilterGroup} from './blocks/filterGroup'
 import octoClient, {OctoClient} from './octoClient'
-import charmClient from '../../charmClient'
+import charmClient from '../../charmFocalBoardClient'
 import {OctoUtils} from './octoUtils'
 import undoManager from './undomanager'
 import {Utils, IDType} from './utils'
