+++ package.json
@@ -46,6 +46,7 @@
     "@mui/material": "^5.2.7",
     "@mui/system": "^5.3.0",
     "@next/bundle-analyzer": "^12.0.7",
+    "@premieroctet/next-crud": "^2.1.4",
     "@prisma/client": "^3.9.2",
     "@reduxjs/toolkit": "^1.6.0",
     "@types/node": "^17.0.13",
