+++ pages/api/[...nextcrud].ts
@@ -1,0 +1,10 @@
+import NextCrud, { PrismaAdapter } from '@premieroctet/next-crud';
+import { prisma } from 'db';
+
+const handler = NextCrud({
+  adapter: new PrismaAdapter({
+    prismaClient: prisma
+  })
+});
+
+export default handler;
