+++ prisma/schema.prisma
@@ -14,10 +14,10 @@ model Space {
   id          String      @id @default(uuid()) @db.Uuid
   deletedAt   DateTime?
   createdAt   DateTime    @default(now())
-  author      User        @relation(fields: [createdBy], references: [id], onDelete: Cascade)
-  createdBy   String
+  author      User        @relation(fields: [createdBy], references: [id], onDelete: Cascade, onUpdate: Cascade)
+  createdBy   String      @db.Uuid
   updatedAt   DateTime
-  updatedBy   String
+  updatedBy   String      @db.Uuid
   name        String
   domain      String      @unique
   permissions SpaceRole[]
@@ -30,7 +30,7 @@ model Page {
   deletedAt   DateTime?
   createdAt   DateTime         @default(now())
   createdBy   String           @db.Uuid
-  author      User             @relation(fields: [createdBy], references: [id], onDelete: Cascade)
+  author      User             @relation(fields: [createdBy], references: [id], onDelete: Cascade, onUpdate: Cascade)
   updatedAt   DateTime
   updatedBy   String           @db.Uuid
   title       String
@@ -40,11 +40,11 @@ model Page {
   icon        String?
   isPublic    Boolean
   path        String
-  parentId    String?
-  spaceId     String?
-  space       Space?           @relation(fields: [spaceId], references: [id], onDelete: Cascade)
+  parentId    String?          @db.Uuid
+  spaceId     String?          @db.Uuid
+  space       Space?           @relation(fields: [spaceId], references: [id], onDelete: Cascade, onUpdate: Cascade)
   type        String
-  boardId     String?
+  boardId     String?          @db.Uuid
   permissions PagePermission[]
   favorites   FavoritePage[]
 }
@@ -58,10 +58,10 @@ model Block {
   User      User      @relation(fields: [createdBy], references: [id])
   updatedAt DateTime
   updatedBy String    @db.Uuid
-  spaceId   String
+  spaceId   String    @db.Uuid
   space     Space     @relation(fields: [spaceId], references: [id], onDelete: Cascade)
   parentId  String
-  rootId    String
+  rootId    String    @db.Uuid
   schema    Int
   type      String
   title     String
@@ -73,7 +73,7 @@ model User {
   createdAt       DateTime         @default(now())
   updatedAt       DateTime?
   addresses       String[]
-  discordId       String?
+  discordId       String?          @unique
   spaceRoles      SpaceRole[]
   favorites       FavoritePage[]
   pagePermissions PagePermission[]
@@ -86,7 +86,7 @@ model FavoritePage {
   pageId String @db.Uuid
   page   Page   @relation(fields: [pageId], references: [id], onDelete: Cascade)
   userId String @db.Uuid
-  user   User   @relation(fields: [userId], references: [id], onDelete: Cascade)
+  user   User   @relation(fields: [userId], references: [id], onDelete: Cascade, onUpdate: Cascade)
 
   @@unique([pageId, userId])
 }
@@ -96,7 +96,7 @@ model PagePermission {
   pageId String @db.Uuid
   page   Page   @relation(fields: [pageId], references: [id], onDelete: Cascade)
   userId String @db.Uuid
-  user   User   @relation(fields: [userId], references: [id], onDelete: Cascade)
+  user   User   @relation(fields: [userId], references: [id], onDelete: Cascade, onUpdate: Cascade)
 
   @@unique([pageId, userId])
 }
@@ -106,7 +106,7 @@ model SpaceRole {
   spaceId String @db.Uuid
   space   Space  @relation(fields: [spaceId], references: [id], onDelete: Cascade)
   userId  String @db.Uuid
-  user    User   @relation(fields: [userId], references: [id], onDelete: Cascade)
+  user    User   @relation(fields: [userId], references: [id], onDelete: Cascade, onUpdate: Cascade)
 
   @@unique([spaceId, userId])
 }
